server/requirements.txt
Flask==0.8
pymongo==2.0.1
pyserial
